Fixes for Rogue Quest
Changes self.inside.type to self.inside.objecttype when Sventer receives a boat type item, and switched to using ITEM_BOAT macro instead of "22" to match on it.  Destroyed items given to the Rogue that aren't on the quest items list to avoid the rogue being overburdened by non-boat items.