Convert connection ip addresse to ipv4 if possible
CServerConfiguration::FromLAN() currently expects
a dotted ipv4 address string only. In
echo_server.cpp:on_message() boost::asio is returning
an ip::address which is generic that holds an ipv4
address mapped to ipv6 space because mplex is
listening on a ipv6 interface.

If the address is mappable back to ipv4 return a ipv4
dotted string or return 0.0.0.0