Bash skill update
Couple of changes
Allows larger range of enemies to be affected depending on stats and
skills - changed from a static 125 max difference between weights to
str+bash skill + 100 and also added an extra that would allow going
beyond that meant for the master bashing trainer to set.

Note that this uses a in self.extra check which I'm not sure about the
efficiency of. Suggestions?

Also added a "spiked" tag for shields to allow some shields to do damage
when they are bashed with. Again, this uses the in operator to check
extras on the shield. May need a better way to handle this.