Spell changes for boosting spells, Re-enable morph
Changes boost spells to allow benefit to having higher amounts of skill
or spell power when casting.
New calculation is hm := hm/10+1;
      if (hm>10)
    hm := 10+((hm-10)/5)
This keeps the old calculation for lower level casting while allowing
benefit at a massively reduced rate for those who are highly skilled and
magically/divinely trained.
This also affects duration whic has been switched to hm*2. This means in
a few cases you will need to cast more often but most of time time it
will be a positive change.
Morphg also re-enabled. Most of the reasons it was disabled are no
longer relevant to the mud in its current state.