Obey function changes
Changed to using a list of allowed commands instead of a list of
forbidden ones.
Changed how the undead command function works to allow invalid commands
to be blocked again.
Changed how whether or not an undead creature is a caster to an extra on
the mob instead of a nonexistent dil function.
Commented out obey function on papi_kitt as it was unused as far as I
can tell.