Updated shield code to keep more of the original code present and add more details
re-adds the shield code including support for a tower shield with different blocking stat requirements (min of str, con, and dex) but a higher base shield bonus.
Makes the blindness check work in both directions instead of just one.
Makes how well the defender succeeded matter for how much damage can be blocked
Encompassed most of the calculations to happen only if the defender is actually using a shield